1. view marks bug fix 2. Added exam vertical report 3. change int to float 4. path change 5. added 2025/26 option in jsp

+++ src/main/java/org/ideoholic/curium/model/teachersperformance/service/TeacherPerformanceService.java
@@ -61,7 +61,7 @@ public class TeacherPerformanceService {
 			
 			SubjectAverage subjectAverage = new SubjectAverage();
 			List<String> classsection = new LinkedList<String>();
-			List<Integer> averageMarksScored = new LinkedList<Integer>();
+			List<Float> averageMarksScored = new LinkedList<Float>();
 			subjectAverage.setExamName("\""+exams.getExamname()+"\"");
 			
 			for (String classOne : classsec) {
@@ -82,8 +82,8 @@ public class TeacherPerformanceService {
 					studentIds.add(student.getStudent().getSid());
 				}
 				List<Marks> marksList = new  MarksDetailsDAO().readListOfMarksPerSubject(studentIds,Integer.parseInt(subject[0]),exams.getExid());
-				int averageMarks = 0;
-				int sum = 0;
+				float averageMarks = 0;
+				float sum = 0;
 				int i = marksList.size();
 				for (Marks marks : marksList) {
